changed the file structure to include the entire project inside blog_api folder
