+++ blog_api/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+
+def create_app():
+    app = Flask(__name__)
+    app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
+    db.init_app(app)
+    return app
