+++ blog_api/blueprints/__init__.py
@@ -1,0 +1,1 @@
+from blog_api.blueprints.user import user
