+++ blog_api/blueprints/user/__init__.py
@@ -1,0 +1,1 @@
+from blog_api.blueprints.user.views import user
