+++ blog_api/blueprints/user/views.py
@@ -1,0 +1,6 @@
+import sys
+
+from flask import Blueprint
+
+print("app" in sys.modules)
+user = Blueprint("user", __name__)
