+++ blog_api/requirements.txt
@@ -1,0 +1,9 @@
+click==8.1.3
+Flask==2.2.2
+Flask-SQLAlchemy==3.0.2
+greenlet==1.1.3.post0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.1
+SQLAlchemy==1.4.42
+Werkzeug==2.2.2
