+++ wsgi.py
@@ -1,16 +1,5 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
+from blog_api import create_app
 
-db = SQLAlchemy()
-
-
-def create_app():
-    app = Flask(__name__)
-    app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
-    db.init_app(app)
-    return app
-
-
-if __name__ == '__main__':
-    _app = create_app()
-    _app.run(debug=True)
+if __name__ == "__main__":
+    app = create_app()
+    app.run(debug=True)
